Jacob_Truong/requirements.txt
google-genai==1.30.0
googlemaps==4.10.0
beautifulsoup4==4.13.4
python-dotenv==1.1.1